pyproject.toml
[project]
name = "pokemon-go"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-bcrypt>=1.0.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.11",
]
